backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdf2image==1.17.0
PyPDF2>=3.0.0
pytesseract==0.3.10
pillow==10.1.0
python-dotenv==1.0.0
google-genai>=0.1.0
python-decouple==3.8
requests>=2.28.0
starlette>=0.27.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
